+++ apps/catalogue/app/stores/useDatasetStore.ts
@@ -3,6 +3,11 @@ import { defineStore } from "pinia";
 import { reactive, ref } from "vue";
 import type { IResources } from "../../interfaces/catalogue";
 import type { ISetting } from "../../../metadata-utils/src";
+import {
+  getHumanReadableString,
+  handleV3Error,
+  toNegotiatorFormat,
+} from "./util/datasetStoreUtils";
 
 export const useDatasetStore = defineStore("datasets", () => {
   const datasets = reactive<Record<string, IResources>>({});
@@ -24,7 +29,6 @@ export const useDatasetStore = defineStore("datasets", () => {
     ]);
 
     const settings = response.data._settings;
-
     const isEnabledSetting = findSetting(CATALOGUE_STORE_IS_ENABLED, settings);
 
     if (isEnabledSetting) {
@@ -65,6 +69,22 @@ export const useDatasetStore = defineStore("datasets", () => {
     }
   }
 
+  async function doStoreRequest() {
+    if (!Object.keys(datasets).length) {
+      return;
+    }
+
+    switch (storeVersion.value) {
+      case "REMS":
+        window.open(datasetStoreUrl.value, "_blank");
+        break;
+      case "negotiatorV3":
+        return await doNegotiatorV3Request();
+      default:
+        return "Unknown data store version, cannot send to store.";
+    }
+  }
+
   async function doNegotiatorV3Request() {
     const url = window.location.origin;
     const humanReadable = getHumanReadableString(datasets);
@@ -86,81 +106,10 @@ export const useDatasetStore = defineStore("datasets", () => {
     } else {
       return handleV3Error(response);
     }
-
-    async function handleV3Error(response: Response) {
-      const statusCode = response.status;
-      const jsonResponse = await response.json();
-      const detail = jsonResponse.detail
-        ? ` Detail: ${jsonResponse.detail}`
-        : "";
-      switch (statusCode) {
-        case 400:
-          const error400 = `Negotiator responded with code 400, invalid input.${detail}`;
-          console.error(error400);
-          return error400;
-        case 401:
-          const error401 = `Negotiator responded with code 401, not authorised.${detail}`;
-          console.error(error401);
-          return error401;
-        case 404:
-          const error404 = `Negotiator not found, error code 404.${detail}`;
-          console.error(error404);
-          return error404;
-        case 413:
-          const error413 = `Negotiator responded with code 413, request too large.${detail}`;
-          console.error(error413);
-          return error413;
-        case 500:
-          const error500 = `Negotiator responded with code 500, internal server error.${detail}`;
-          console.error(error500);
-          return error500;
-        default:
-          const errorUnknown = `An unknown error occurred with the Negotiator. Please try again later.${detail}`;
-          console.error(errorUnknown);
-          return errorUnknown;
-      }
-    }
-
-    function toNegotiatorFormat(datasets: Record<string, IResources>) {
-      return Object.values(datasets).map((dataset) => ({
-        id: dataset.pid,
-        name: dataset.name,
-      }));
-    }
-
-    function getHumanReadableString(datasets: Record<string, IResources>) {
-      const datasetInfo = Object.values(datasets).map((dataset) => {
-        return { pid: dataset.pid, name: dataset.name };
-      });
-      const humanReadableString = datasetInfo
-        .reduce((acc, dataset) => {
-          return acc + `${dataset.name} (${dataset.pid}), `;
-        }, "")
-        .slice(0, -2);
-
-      return humanReadableString;
-    }
-  }
-
-  async function doStoreRequest() {
-    if (!Object.keys(datasets).length) {
-      return;
-    }
-
-    switch (storeVersion.value) {
-      case "REMS":
-        window.open(datasetStoreUrl.value, "_blank");
-        break;
-      case "negotiatorV3":
-        return await doNegotiatorV3Request();
-      default:
-        return "Unknown data store version, cannot send to store.";
-    }
   }
 
   return {
     datasets,
-    datasetStoreUrl,
     isEnabled,
     storeVersion,
     addToCart,
